requirements.txt
PyGithub>=1.32
pykwalify>=1.5.0
ruamel.yaml>=0.10.0
